+++ examples/flutter_embedding/lib/main.dart
@@ -25,6 +25,8 @@ void main() {
 
       link(rel: 'manifest', href: 'manifest.json'),
       link(rel: 'stylesheet', href: 'css/style.css', type: 'text/css'),
+
+      script(src: "flutter_bootstrap.js", async: true, []),
     ],
     body: App(),
   ));