+++ packages/jaspr_flutter_embed/lib/src/view_constraints/view_constraints_io.dart
@@ -1,0 +1,8 @@
+class ViewConstraints {
+  ViewConstraints({this.minWidth, this.maxWidth, this.minHeight, this.maxHeight});
+
+  final double? minWidth;
+  final double? maxWidth;
+  final double? minHeight;
+  final double? maxHeight;
+}