+++ packages/jaspr_flutter_embed/lib/src/view_constraints/view_constraints_web.dart
@@ -1,0 +1,10 @@
+import 'dart:js_interop';
+
+extension type ViewConstraints._(JSObject _) {
+  external ViewConstraints({double? minWidth, double? maxWidth, double? minHeight, double? maxHeight});
+
+  external double? minWidth;
+  external double? maxWidth;
+  external double? minHeight;
+  external double? maxHeight;
+}